requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pymysql>=1.1,<2.0
python-dotenv>=1.0,<2.0
google-auth>=2.32,<3.0
email-validator>=2.2,<3.0
requests
itsdangerous